Moved JoiningCommunityController to the controller/joining package. Added comments.
