+++ practiceapp/src/main/java/com/practiceapp/practiceapp/entity/CommunityEntity.java
@@ -46,6 +46,10 @@ public class CommunityEntity {
 
     private List<String> topics = new ArrayList<>();
 
+    //@JsonBackReference annotation here is to resolve the infinite loop issue (which ends up in stack overflow error) when trying to load a user or a community
+    // The reasoning is that, when trying to load a user, it tries to load the dbref for the community in its joined communities list, and when trying to load that community,
+    // it tries to load the dbref for the user in its members list. This goes on forever and ends up in an infinite loop.
+
     // References to users and posts:
     @JsonBackReference
     @DBRef(lazy = true)
