+++ practiceapp/src/main/java/com/practiceapp/practiceapp/entity/UserEntity.java
@@ -32,6 +32,10 @@ public class UserEntity {
     @Field
     private String email;
 
+    //@JsonManagedReference annotation here is to resolve the infinite loop issue (which ends up in stack overflow error) when trying to load a user or a community
+    // The reasoning is that, when trying to load a user, it tries to load the dbref for the community in its joined communities list, and when trying to load that community,
+    // it tries to load the dbref for the user in its members list. This goes on forever and ends up in an infinite loop.
+
     @JsonManagedReference
     @DBRef(lazy = true)
     private List<CommunityEntity> joined_communities = new ArrayList<>();
